feat: Add login and profile page functionality
The code changes include adding JavaScript functionality for the login page and profile page. The login page now prevents the form from submitting the traditional way and sends an AJAX request to authenticate the user. If the authentication is successful, the user is redirected to the profile page. If there is an error, an error message is displayed.

For the profile page, the JavaScript code populates the user's profile data and allows the user to edit and save their profile information. The profile data is retrieved using an AJAX request with the user's token. The user can click on the "Update" button to enable editing of the input fields and select dropdowns. The "Save" button sends a PUT request to update the user's profile data.

These changes are in line with the recent user commits and repository commits that focused on updating CSS styles and adding HTML templates for the login and exam pages.