feat: Update URLs for API endpoints
This commit updates the URLs for the API endpoints to use the production server instead of the local development server. The base URL for the accounts API is changed to 'https://maat-system.s1cario.tech/api/accounts/', and the base URL for the exams API is changed to 'https://maat-system.s1cario.tech/api/exams/'. This change ensures that the correct URLs are used for accessing the accounts and exams data in the production environment.